refactor: put render function in Mark and removed it from Point
