+++ src/mixins/Marks/Mark.js
@@ -1,6 +1,8 @@
 import CoordinateTreeUser from '../CoordinateTreeUser.js'
 import DataReceiver from '../Data/DataReceiver.js'
 
+import mapAesthetics from '../../components/Marks/utils/mapAesthetics.js'
+
 import {
   parseAesthetic,
   parseCoordinate,
@@ -56,5 +58,29 @@ export default {
     this.parseCoordinateSet.bind(this)
     this.parseGeometry.bind(this)
     this.parseProperty.bind(this)
+  },
+
+  render (createElement) {
+    if (this.__update) {
+      if (!this.$$map) {
+        // Create svg element using aesthetics
+        return this.renderSVG(createElement, this.aesthetics)
+      }
+
+      if (this.$$map) {
+        // Create the aesthetics for each mark
+        let aestheticsPerMark = mapAesthetics(this.aesthetics, this.context)
+
+        // Create svg element for each mark from aesthetics
+        let components = []
+        for (let aesthetics of aestheticsPerMark) {
+          components.push(
+            this.renderSVG(createElement, aesthetics)
+          )
+        }
+
+        return createElement('g', components)
+      }
+    }
   }
 }
